fix: Restructure files. Clean up, remove dead code
