+++ v3/src/components/graph/adornments/movable-line/movable-line-types.ts
@@ -1,5 +1,4 @@
 export const kMovableLineClass = "movable-line"
 export const kMovableLineType = "Movable Line"
 export const kMovableLinePrefix = "ADRN"
-export const kMovableLineLabel = "Movable Line"
-export const kMovableLineHint = ""
+export const kMovableLineLabelKey = "DG.Inspector.graphMovableLine"
