+++ v3/src/components/graph/adornments/movable-point/movable-point-registration.ts
@@ -1,0 +1,21 @@
+import { registerAdornmentComponentInfo } from "../adornment-component-info"
+import { registerAdornmentContentInfo } from "../adornment-content-info"
+import { MovablePointModel } from "../adornment-models"
+import { kMovablePointClass, kMovablePointLabelKey, kMovablePointPrefix,
+         kMovablePointType } from "./movable-point-types"
+import { MovablePoint } from "./movable-point"
+
+registerAdornmentContentInfo({
+  type: kMovablePointType,
+  plots: ['scatterPlot'],
+  prefix: kMovablePointPrefix,
+  modelClass: MovablePointModel
+})
+
+registerAdornmentComponentInfo({
+  adornmentEltClass: kMovablePointClass,
+  Component: MovablePoint,
+  labelKey: kMovablePointLabelKey,
+  order: 10,
+  type: kMovablePointType
+})
