+++ v3/src/components/graph/adornments/movable-point/movable-point-types.ts
@@ -1,5 +1,5 @@
 export const kMovablePointClass = "movable-point"
 export const kMovablePointType = "Movable Point"
 export const kMovablePointPrefix = "ADRN"
-export const kMovablePointLabel = "Movable Point"
+export const kMovablePointLabelKey = "DG.Inspector.graphMovablePoint"
 export const kMovablePointHint = ""
