+++ v3/src/components/graph/adornments/movable-point/movable-point.tsx
@@ -31,8 +31,8 @@ export const MovablePoint = observer(function MovablePoint(props: {
     layout = useAxisLayoutContext(),
     xScale = layout.getAxisScale("bottom") as ScaleNumericBaseType,
     yScale = layout.getAxisScale("left") as ScaleNumericBaseType,
-    graphHeight = layout.getAxisLength('left') ?? 0,
-    graphWidth = layout.getAxisLength('bottom') ?? 0,
+    graphHeight = layout.getAxisLength('left'),
+    graphWidth = layout.getAxisLength('bottom'),
     xSubAxesCount = layout.getAxisMultiScale('bottom')?.repetitions ?? 1,
     ySubAxesCount = layout.getAxisMultiScale('left')?.repetitions ?? 1,
     xDomain = xAxis.domain,
