+++ v3/src/components/graph/adornments/movable-value/movable-value.tsx
