+++ v3/src/components/graph/graphing-types.ts
@@ -88,5 +88,3 @@ export const kGraphClass = "graph-plot"
 export const kGraphClassSelector = `.${kGraphClass}`
 export const kGraphAdornmentsClass = "graph-adornments-grid"
 export const kGraphAdornmentsClassSelector = `.${kGraphAdornmentsClass}`
-export const kMovableLineType = "Movable Line"
-export const kMovablePointType = "Movable Point"
